MIGRATE:FRESH NECESARIOgit add . Seguridad prácticamente completado, falta implementación en algunos módulos y otras cosas

+++ app/Http/Controllers/Perfil/PerfilController.php
@@ -4,11 +4,36 @@ namespace App\Http\Controllers\Perfil;
 
 use App\Perfil;
 use App\Modulo;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
 class PerfilController extends Controller
 {
+    const PERFIL_ID_ADMIN = 1;
+
+    public function __construct() {
+        $this->middleware(function ($request, $next) {
+            if(Auth::check()) {
+                $user = Auth::user();
+                $modulos = $user->perfil->modulos;
+                foreach ($modulos as $modulo) {
+                    if($modulo->nombre == "seguridad")
+                        return $next($request);
+                }
+                return redirect()->route('denegado');
+            } else
+                return redirect()->route('login');
+        });
+    }
+
+    public function hasSecurity($perfil) {
+        foreach ($perfil->modulos as $modulo)
+            if($modulo->nombre == 'seguridad')
+                return true;
+        return false;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -39,11 +64,26 @@ class PerfilController extends Controller
      */
     public function store(Request $request)
     {
-        $perfil = Perfil::create($request->all());
-        $modulos = Modulo::find($request->input('modulo_id'));
-        $perfil->modulos()->attach($modulos);
-        $modulos = Modulo::get();
-        return view('seguridad.perfil.view', ['perfil' => $perfil, 'modulos' => $modulos]);
+        $seguridad = false;
+        if($request->input('modulo_id') != null)
+            foreach ($request->input('modulo_id') as $id)
+                if(Modulo::find($id)->nombre == 'seguridad')
+                    $seguridad = true;
+
+        if(Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad)
+            return redirect()->route('denegado');
+        else {
+            $rules = [
+                'nombre'=>'required|string|unique:perfils',
+                'modulo_id'=>'nullable|array'
+            ];
+            $this->validate($request, $rules);
+            $perfil = Perfil::create($request->all());
+            $modulos = Modulo::find($request->input('modulo_id'));
+            $perfil->modulos()->attach($modulos);
+            $modulos = Modulo::get();
+            return view('seguridad.perfil.view', ['perfil' => $perfil, 'modulos' => $modulos]);
+        }
     }
 
     /**
@@ -55,8 +95,14 @@ class PerfilController extends Controller
     public function show($id)
     {
         $perfil = Perfil::find($id);
-        $modulos = Modulo::get();
-        return view('seguridad.perfil.view', ['perfil' => $perfil, 'modulos' => $modulos]);
+        $seguridad = $this->hasSecurity($perfil);
+
+        if($perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $modulos = Modulo::get();
+            return view('seguridad.perfil.view', ['perfil' => $perfil, 'modulos' => $modulos]);
+        }
     }
 
     /**
@@ -68,8 +114,14 @@ class PerfilController extends Controller
     public function edit($id)
     {
         $perfil = Perfil::find($id);
-        $modulos = Modulo::get();
-        return view('seguridad.perfil.edit', ['perfil' => $perfil, 'modulos' => $modulos]);
+        $seguridad = $this->hasSecurity($perfil);
+
+        if($perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $modulos = Modulo::get();
+            return view('seguridad.perfil.edit', ['perfil' => $perfil, 'modulos' => $modulos]);
+        }
     }
 
     /**
@@ -82,12 +134,27 @@ class PerfilController extends Controller
     public function update(Request $request, $id)
     {
         $perfil = Perfil::find($id);
-        $perfil->update($request->all());
-        $modulos = Modulo::get();
-        $mods = Modulo::find($request->input('modulo_id'));
-        $perfil->modulos()->detach();
-        $perfil->modulos()->attach($mods);
-        return view('seguridad.perfil.view', ['perfil' => $perfil, 'modulos' => $modulos]);
+        $seguridad = false;
+        if($request->input('modulo_id') != null)
+            foreach ($request->input('modulo_id') as $id)
+                if(Modulo::find($id)->nombre == 'seguridad')
+                    $seguridad = true;
+
+        if($perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $rules = [
+                'nombre'=>'required|string|unique:perfils',
+                'modulo_id'=>'nullable|array'
+            ];
+            $this->validate($request, $rules);
+            $perfil->update($request->all());
+            $modulos = Modulo::get();
+            $mods = Modulo::find($request->input('modulo_id'));
+            $perfil->modulos()->detach();
+            $perfil->modulos()->attach($mods);
+            return view('seguridad.perfil.view', ['perfil' => $perfil, 'modulos' => $modulos]);
+        }
     }
 
     /**
@@ -99,9 +166,15 @@ class PerfilController extends Controller
     public function destroy($id)
     {
         $perfil = Perfil::find($id);
-        $perfil->modulos()->detach();
-        $perfil->delete();
-        $perfiles = Perfil::get();
-        return view('seguridad.perfil.index', ['perfiles' => $perfiles]);
+        $seguridad = $this->hasSecurity($perfil);
+
+        if($perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $perfil->modulos()->detach();
+            $perfil->delete();
+            $perfiles = Perfil::get();
+            return view('seguridad.perfil.index', ['perfiles' => $perfiles]);
+        }
     }
 }
