+++ app/Http/Controllers/Usuario/UsuarioController.php
@@ -2,28 +2,69 @@
 
 namespace App\Http\Controllers\Usuario;
 
-use App\Usuario;
+use App\User;
 use App\Perfil;
 use App\Puesto;
 use App\Area;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
 
 class UsuarioController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Constante del id del perfil del admin
+     *
+     * @default 1
+     */
+    const PERFIL_ID_ADMIN = 1;
+
+    /**
+     * Usa un middleware para validar la sesión y el acceso al módulo de seguridad
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function __construct() {
+        $this->middleware(function ($request, $next) {
+            if(Auth::check()) {
+                $user = Auth::user();
+                $modulos = $user->perfil->modulos;
+                foreach ($modulos as $modulo) {
+                    if($modulo->nombre == "seguridad")
+                        return $next($request);
+                }
+                return redirect()->route('denegado');
+            } else
+                return redirect()->route('login');
+        });
+    }
+
+    /**
+     * Verifica si un perfil contiene al módulo de seguridad
+     *
+     * @param App\Perfil $perfil
+     * @return boolean
+     */
+    public function hasSecurity($perfil) {
+        foreach ($perfil->modulos as $modulo)
+            if($modulo->nombre == 'seguridad')
+                return true;
+        return false;
+    }
+
+    /**
+     * Muestra un listado de los usuarios registrados
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        $usuarios = Usuario::get();
+        $usuarios = User::get();
         return view('seguridad.usuario.index', ['usuarios' => $usuarios]);
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Muestra el formulario de registro de usuarios
      *
      * @return \Illuminate\Http\Response
      */
@@ -36,46 +77,87 @@ class UsuarioController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Almacena un nuevo usuario en la base de datos
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $usuario = Usuario::create($request->all());
-        return view('seguridad.usuario.view', ['usuario' => $usuario]);
+        $seguridad = $this->hasSecurity(Perfil::find($request->input('perfil_id')));
+
+        if($request->input('perfil_id') == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $rules = [
+                'perfil_id'=>'required|integer',
+                'puesto_id'=>'required|integer',
+                'area_id'=>'required|integer',
+                'name'=>'required|alpha',
+                'email' => 'required|string|email|unique:users',
+                'password' => 'required|string',
+                'nombre'=>'required|alpha',
+                'appaterno'=>'required|alpha',
+                'apmaterno'=>"nullable|alpha"
+            ];
+            $this->validate($request, $rules);
+            $inputs = $request->all();
+            $usuario = User::create([
+                'name'=> $inputs['name'],
+                'email'=>$inputs['email'],
+                'password'=>bcrypt($inputs['password']),
+                'nombre'=>$inputs['nombre'],
+                'appaterno'=>$inputs['appaterno'],
+                'apmaterno'=>$inputs['apmaterno'],
+                'area_id'=>$inputs['area_id'],
+                'perfil_id'=>$inputs['perfil_id'],
+                'puesto_id'=>$inputs['puesto_id'],
+            ]);
+            return view('seguridad.usuario.view', ['usuario' => $usuario]);
+        }
     }
 
     /**
-     * Display the specified resource.
+     * Muestra a un usuario en concreto
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
-        $usuario = Usuario::find($id);
-        return view('seguridad.usuario.view', ['usuario' => $usuario]);
+        $usuario = User::find($id);
+        $seguridad = $this->hasSecurity($usuario->perfil);
+
+        if($usuario->perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            return view('seguridad.usuario.view', ['usuario' => $usuario]);
+        }
     }
 
     /**
-     * Show the form for editing the specified resource.
+     * Muestra el formulario para editar a un usuario en concreto
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
-        $perfiles = Perfil::get();
-        $puestos = Puesto::get();
-        $areas = Area::get();
-        $usuario = Usuario::find($id);
-        return view('seguridad.usuario.edit', ['usuario' => $usuario, 'perfiles' => $perfiles, 'puestos' => $puestos, 'areas' => $areas]);
+        $usuario = User::find($id);
+        $seguridad = $this->hasSecurity($usuario->perfil);
+
+        if($usuario->perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $perfiles = Perfil::get();
+            $puestos = Puesto::get();
+            $areas = Area::get();
+            return view('seguridad.usuario.edit', ['usuario' => $usuario, 'perfiles' => $perfiles, 'puestos' => $puestos, 'areas' => $areas]);
+        }
     }
 
     /**
-     * Update the specified resource in storage.
+     * Actualiza los datos de un usuario en concreto en la base de datos
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -83,21 +165,55 @@ class UsuarioController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $usuario = Usuario::find($id);
-        $usuario->update($request->all());
-        return view('seguridad.usuario.view', ['usuario' => $usuario]);
+        $usuario = User::find($id);
+        $seguridad = $this->hasSecurity($usuario->perfil);
+
+        if($request->input('perfil_id') == self::PERFIL_ID_ADMIN || $usuario->perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else {
+            $rules = [
+                'perfil_id'=>'required|integer',
+                'puesto_id'=>'required|integer',
+                'area_id'=>'required|integer',
+                'name'=>'required|',
+                'email' => 'required|string|email',
+                'password' => 'nullable|string',
+                'nombre'=>'required|alpha',
+                'appaterno'=>'required|alpha',
+                'apmaterno'=>"nullable|alpha"
+            ];
+            $this->validate($request, $rules);
+            $usuario->name = $request->input('name');
+            $usuario->email = $request->input('email');
+            if($request->input('password') != null)
+                $usuario->password = bcrypt($request->input('password'));
+            $usuario->nombre = $request->input('nombre');
+            $usuario->appaterno = $request->input('appaterno');
+            $usuario->apmaterno = $request->input('apmaterno');
+            $usuario->area_id = $request->input('area_id');
+            $usuario->perfil_id = $request->input('perfil_id');
+            $usuario->puesto_id = $request->input('puesto_id');
+            $usuario->save();
+            $usuario = $usuario->fresh();
+            return view('seguridad.usuario.view', ['usuario' => $usuario]);
+        }
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Elimina a un usuario en concreto de la base de datos
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $usuario = Usuario::find($id);
-        $usuario->delete();
-        return $this->index();
+        $usuario = User::find($id);
+        $seguridad = $this->hasSecurity($usuario->perfil);
+
+        if($usuario->perfil->id == self::PERFIL_ID_ADMIN || (Auth::user()->perfil->id != self::PERFIL_ID_ADMIN && $seguridad))
+            return redirect()->route('denegado');
+        else
+            $usuario->delete();
+            return $this->index();
     }
 }
